GETP-216 refactor: persistence 모듈의 엔티티가 존재하지 않은 경우의 예외들이 NotFoundException을 상속하도록 변경
